Created ImportGenerator which can import the existing graph and recreate the associated program objects as if the program had never been terminated.

+++ src/graphgame/cell/Cell.java
@@ -3,7 +3,7 @@ package graphgame.cell;
 import java.util.HashMap;
 
 public interface Cell {
-    long getNodeID();
+    int getNodeID();
 
     default boolean willAssociate(Cell cell){
         return true;
